% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma.r
\name{exclusions_digraph}
\alias{exclusions_digraph}
\title{Generic exclusions flow chart (beta)}
\usage{
exclusions_digraph(incl_counts, total_label = "Total",
  incl_labels = LETTERS[seq_len(length(incl_counts) - 1)],
  excl_labels = paste("Not ", incl_labels), percent_of_total = FALSE,
  percent_of_prev = FALSE, font_size = 10, dpi = 72, ...)
}
\arguments{
\item{...}{arguments passed to \code{scales::percent}}
}
\description{
Generate the diagraph DOT description for a generic series of inclusions and exclusions
}
\examples{
PRISMAstatement:::exclusions_digraph(c(1000, 300, 150, 75, 38))
dot <-
  PRISMAstatement:::exclusions_digraph(
    incl_counts = c(972, 132, 77, 14),
    total_label = "Total Screened",
    incl_labels = c("Consented", "Completed Study", "BMI <= 30"),
    excl_labels = c("Declined Consent", "Failed to Complete", "BMI > 30")
    )
DiagrammeR::grViz(dot)
}
\keyword{internal}
