% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma.r
\name{prisma_pdf}
\alias{prisma_pdf}
\title{Make PDF of the plot}
\usage{
prisma_pdf(x, filename = "prisma.pdf")
}
\arguments{
\item{x}{output of call to \code{prisma}}

\item{filename}{path of output file}
}
\description{
This makes a PDF file which can be included by knitr Sweave.
}
\examples{
\dontrun{
g <- prisma(9, 8, 7, 6, 5, 4, 3, 2, 1)
prisma_pdf(g, "test.pdf")
knitr::include_graphics("test.pdf")
}
}
\keyword{internal}
