% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma.r
\name{flow_exclusions}
\alias{flow_exclusions}
\alias{flow_exclusions_dot}
\title{Generic exclusions flow chart (beta)}
\usage{
flow_exclusions(incl_counts, total_label = "Total",
  incl_labels = LETTERS[seq_len(length(incl_counts) - 1)],
  excl_labels = paste("Not ", incl_labels), show_count = TRUE,
  percent_of_total = FALSE, percent_of_prev = FALSE, font_size = 10,
  dpi = 72, ...)

flow_exclusions_dot(incl_counts, total_label = "Total",
  incl_labels = LETTERS[seq_len(length(incl_counts) - 1)],
  excl_labels = paste("Not ", incl_labels), show_count = TRUE,
  percent_of_total = FALSE, percent_of_prev = FALSE, font_size = 10,
  dpi = 72, ...)
}
\arguments{
\item{...}{arguments passed to \code{scales::percent} and \code{DiagrammeR::grViz}}
}
\description{
Generate the diagraph DOT description for a generic series of inclusions and
exclusions
}
\section{Functions}{
\itemize{
\item \code{flow_exclusions_dot}: Just generate the DOT text
}}

\examples{
flow_exclusions(c(1000, 300, 150, 75, 38))
dot <-
  flow_exclusions_dot(
    incl_counts = c(972, 132, 77, 14),
    total_label = "Total Screened",
    incl_labels = c("Consented", "Completed Study", "BMI <= 30"),
    excl_labels = c("Declined Consent", "Failed to Complete", "BMI > 30")
    )
DiagrammeR::grViz(dot)
flow_exclusions(c(1000, 300, 150, 75, 38), percent_of_total = TRUE)
flow_exclusions(c(100000, 3000, 1666, 411, 38),
                percent_of_prev = TRUE,
                show_count = FALSE)
}
\keyword{internal}
